feat(clone): Add clonePokemon functionality
This commit adds the ability to clone a Pokemon. The `clonePokemon` method in the `EntitiesService` class is implemented to create a new Pokemon object based on an existing one. The cloned Pokemon is then added to the local storage and returned as an observable. Additionally, the necessary actions and effects are added to handle the cloning process in the store. A new component called `PokemonCloneComponent` is also created to display a button for cloning a Pokemon.